# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parts:
    get:
      operationId: listParts
      tags: [parts]
      summary: List all parts
      responses:
        "200":
          description: Parts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartsResponse"
    post:
      operationId: createPart
      tags: [parts]
      summary: Create a part with a permanent public identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartInput"
      responses:
        "201":
          description: Created part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
  /parts/import:
    post:
      operationId: importParts
      tags: [parts]
      summary: Import existing client-side parts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartImportInput"
      responses:
        "200":
          description: Imported and existing parts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartsResponse"
  /parts/{partId}:
    patch:
      operationId: updatePart
      tags: [parts]
      summary: Update a part without changing its public identifier
      parameters:
        - name: partId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartUpdate"
      responses:
        "200":
          description: Updated part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
    delete:
      operationId: deletePart
      tags: [parts]
      summary: Delete a part
      parameters:
        - name: partId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Part deleted
  /public/parts/{publicId}:
    get:
      operationId: getPublicPart
      tags: [parts]
      summary: Get the latest public details for a QR identifier
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9_-]{20,80}$"
      responses:
        "200":
          description: Current part details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
        "404":
          description: Part not found
...
  /donors:
    get:
      operationId: listDonors
      tags: [donors]
      summary: List shared donor vehicles
      responses:
        "200":
          description: Donors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorsResponse"
    post:
      operationId: createDonor
      tags: [donors]
      summary: Create a donor using a caller supplied stable ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorInput"
      responses:
        "201":
          description: Created donor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "409":
          description: ID already exists or was deleted
  /donors/import:
    post:
      operationId: importDonors
      tags: [sync]
      summary: Add missing legacy donors without overwriting server records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorImportInput"
      responses:
        "200":
          description: Current shared donors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorsResponse"
  /donors/{donorId}:
    patch:
      operationId: updateDonor
      tags: [donors]
      summary: Update one donor with optimistic conflict detection
      parameters:
        - name: donorId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorUpdate"
      responses:
        "200":
          description: Updated donor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "404":
          description: Donor not found
        "409":
          description: Version conflict
    delete:
      operationId: deleteDonor
      tags: [donors]
      summary: Delete a donor while preserving associated parts
      parameters:
        - name: donorId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: preserveParts
          in: query
          required: false
          description: Required when parts still reference this donor; parts stay visible as orphaned server records.
          schema:
            type: boolean
      responses:
        "204":
          description: Donor deleted
        "404":
          description: Donor not found
        "409":
          description: Donor has parts and explicit preservation was not requested
...
  /settings/{key}:
    put:
      operationId: upsertSyncSetting
      tags: [settings]
      summary: Create or replace one shared setting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncSettingInput"
      responses:
        "200":
          description: Shared setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncSetting"
        "409":
          description: Version conflict
    delete:
      operationId: deleteSyncSetting
      tags: [settings]
      summary: Delete one shared setting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      responses:
        "204":
          description: Setting deleted
        "404":
          description: Setting not found
components:
  schemas:
...
    Part:
      type: object
      required: [id, publicId, donorId, donorLabel, name, code, price, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        publicId:
          type: string
        donorId:
          type: string
        donorLabel:
          type: string
        name:
          type: string
        code:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [inventory, sold]
        location:
          type: ["string", "null"]
        soldAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PartsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Part"
    PartInput:
      type: object
      required: [donorId, donorLabel, name, code, price]
      properties:
        donorId:
          type: string
          minLength: 1
        donorLabel:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        code:
          type: string
        price:
          type: number
          minimum: 0
        location:
          type: string
    PartUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
        price:
          type: number
          minimum: 0
        status:
          type: string
          enum: [inventory, sold]
        location:
          type: ["string", "null"]
    PartImportItem:
      type: object
      required: [legacyId, donorId, donorLabel, name, code, price, status, createdAt]
      properties:
        legacyId:
          type: string
          minLength: 1
        donorId:
          type: string
          minLength: 1
        donorLabel:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        code:
          type: string
        price:
          type: number
          minimum: 0
        status:
          type: string
          enum: [inventory, sold]
        location:
          type: string
        createdAt:
          type: string
          format: date-time
        soldAt:
          type: string
          format: date-time
    PartImportInput:
      type: object
      required: [parts]
      properties:
        parts:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/PartImportItem"
...
    Donor:
      type: object
      required: [id, year, make, model, engine, fuel, mileage, purchasePrice, status, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          minLength: 1
        year:
          type: integer
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        engine:
          type: string
        fuel:
          type: string
        mileage:
          type: integer
          minimum: 0
        purchasePrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [active, closed]
        location:
          type: ["string", "null"]
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DonorsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Donor"
    DonorInput:
      type: object
      required: [id, year, make, model, engine, fuel, mileage, purchasePrice]
      properties:
        id:
          type: string
          minLength: 1
        year:
          type: integer
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        engine:
          type: string
        fuel:
          type: string
        mileage:
          type: integer
          minimum: 0
        purchasePrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [active, closed]
        location:
          type: string
        createdAt:
          type: string
          format: date-time
    DonorUpdate:
      allOf:
        - $ref: "#/components/schemas/DonorInputPartial"
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion:
              type: integer
              minimum: 1
    DonorInputPartial:
      type: object
      properties:
        year:
          type: integer
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        engine:
          type: string
        fuel:
          type: string
        mileage:
          type: integer
          minimum: 0
        purchasePrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [active, closed]
        location:
          type: ["string", "null"]
    DonorImportInput:
      type: object
      required: [donors]
      properties:
        donors:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/DonorInput"
    SyncSetting:
      type: object
      required: [key, value, version, updatedAt]
      properties:
        key:
          type: string
        value:
          type: object
          additionalProperties: true
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    SyncSettingInput:
      type: object
      required: [value]
      properties:
        value:
          type: object
          additionalProperties: true
        expectedVersion:
          type: integer
          minimum: 1