# lib/api-zod/src/generated/types/donorUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DonorInputPartial } from './donorInputPartial';

export type DonorUpdate = DonorInputPartial & {
  /** @minimum 1 */
  expectedVersion: number;
};