# lib/api-zod/src/generated/types/syncSettingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncSettingInputValue } from './syncSettingInputValue';

export interface SyncSettingInput {
  value: SyncSettingInputValue;
  /** @minimum 1 */
  expectedVersion?: number;
}